WIP: Moved AnalyticalCardHeader to components folder
